Added javaDoc comment to sellShares and buyShares

+++ StockHolding.java
@@ -2,7 +2,7 @@
  * Write a description of class Asset here.
  *
  * @author (Matthew Laframboise, Nate Sletten)
- * @version (a version number or a date)
+ * @version (Spring 2023)
  */
 public class StockHolding
 {
@@ -32,10 +32,23 @@ public class StockHolding
         return price;
     }
     
-    public void buyShares(int numShares, int price){
+    /**
+     * This method updates the instance variables for the number of shares to buy and the price per share. 
+     * @param numShares This is and int value for the number of shares bought.
+     * @param price This is a double value for the price per share bought.
+     */
+    public void buyShares(int numShares, double price){
         this.numShares = numShares;
         this.price = price;
     }
+    
+    /**
+     * This method is used to sell shares. It will first check to make sure the number of shares you want to sell is less the the amount of shares purchased.
+     * If the number of shares you want to sell is less then the total amount of shares owned then it updates the instance variables and returns the dollar
+     * amount. If you try to sell more shares then you own it will return 0.
+     * @param numSell This is the number of shares you want to sell.
+     * @return This returns the amount you made on the sale if you sell the number of shares you own or less. 
+     */
     public double sellShares(int numSell){
         double amount = 0;
         if (numShares - numSell >= 0){
